Add the outlets to the ViewController.
Add myLocation property and set the 2 labels as observers for any value change.

Add a locationManager.

Add the proper delegate methods and set the delegate of the location manager to the ViewController

Add the proper action methods to each of the 4 buttons

Add the didUpdateLocatons methods that sets myLocation property to the last acquired location from the locationManager.

Add the proper permissions to the info.plist file.